Fix potential dangling references in CanrosClient
`CanrosClient` was creating temporary `boost::function`s and passing them to `NodeHandle::advertise` by reference. This seems like it might create dangling references so I updated it to use a `boost::function` stored as a member variable.

As a bonus the observer stuff is split out of `CanrosClient` which cleans up the code a little.