Get CAN working in melodic (#278)
* Get CAN working in melodic

Reverts to uavcan v0, and replaces the old mapper node (which was removed a while back) with a C++ canros client. The uavcan v1 interface is kept in case we need it later.

* Fix potential dangling references in CanrosClient

`CanrosClient` was creating temporary `boost::function`s and passing them to `NodeHandle::advertise` by reference. This seems like it might create dangling references so I updated it to use a `boost::function` stored as a member variable.

As a bonus the observer stuff is split out of `CanrosClient` which cleans up the code a little.

* Clean up a couple things

* Fix formatting

* Add mapper messages again

These are being used by the C++ mapper node. Since I'm keeping that around these have to be kept as well so the code compiles.